Removed the dependency on the Checker class:
In the last commit I started migrating the code to having utilities
that are used in the code. I migrated the rest of the code to use
these utility functions instead of the Checker class. I also changed
the utility function a little to make it compatible with the code.
I changed some tests for this too.