Started adding tests for AI code, refactored AI code a little
1) Refactored AI code a little bit. I decided to change one of the
functions to use a dictionary parameter instead of 3 extra parameters,
set the dictionary to be an instance variable, and set a method to
clean it. I decided to put this into another function that takes an
input function, runs it and then runs this method. This function is
returned and run instead, and this is done using a decorator.

2) I added test code for the AI. I've mainly been testing all the
helper functions, and will then test out scenarios on the main function.
I may add more tests for the existing functions that I've tested for
thoroughness. It'll be a trade-off though, with portability.