Changed the AI algorithm, split up tests for AI into different files:
1) I found a bug in the AI algorithm when looking for forks (2 unblocked
rows of 2 on the board, meaning an automatic win in 2 moves). Depending
on where you are, you will look diagonally across the board, and if you
find one set of 2 in a row (unblocked) diagonally, you don't do it
again. But that becomes a problem if your current coordinate is the
center piece and the board looks like this:

  |  |
  |  |
-- -- --
  |\/|/\
  |/\|\/
-- -- --
\/|/\|\/
/\|\/|/\

The After finding the first unblocked row of 2, the program doesn't look
diagonally because that direction is "turned off", meaning the computer
misses the chance to get 2 unblocked rows of 2 possibly, and win the
game. So now I have diagonal directions of "diagonal_1" and
"diagonal_2". "diagonal_1" represents going from the top-left to the
bottom-right, and "diagonal_2" from top-right to bottom-left. I
"turn off" these directions when I find an unblocked row of 2 there.

2) I found that my previous way of finding possible forks that the
opponent could make was not the best solution. Even now it still isn't.
But the idea is based on the strategy found here:
https://en.wikipedia.org/wiki/Tic-tac-toe#Strategy

If there are multiple possible places on the board where the opponent
can get a fork (multiple "fork spots"), then look where your pieces are
on the board, look where the possible fork sopts are for the opponent,
and find a spot such that you create 2 in a row, and the move to block
it does not create a fork for the opponent. You do this if there are
multiple fork spots on the board. I added a function and adjusted
my minimax algorithm to do this (function called find_optimal_move).

3) I split up the old .py file for all python tests
(tests/test_artificial_intelligence.py) into 2 test files: one for
the helper functions (test_ai_helper_fns) and one for the actual
minimax function. I have added all according tests, and may add a
couple more.

TL;DR fixed bugs, split up the test files into 2 test files