Fixed bug, added alpha beta pruning, added more tests also:
1) In our "find_optimal_move" function, I hadn't taken into account
what happens if no rows are returned where they are empty (only one
piece). This bug showed itself when I was playing the game and it
threw an error. So in the event that no rows are returned, just
return the first element in the "fork_spots" list (a list of all spots
where we can block a fork for the opponent).

2) I added alpha-beta pruning. For testing reasons I made it optional
to use alpha-beta pruning in the minimax algorithm.

3) I added tests for alpha beta pruning, checking to make sure that the
static evaluations of the board are computed less when using Alpha-Beta
pruning than they are without it. This serves to make sure that we save
computation. I also added a test for the minimax algorithm at 4 layers
of recursion