Added functionality for various difficulty levels:
I came up with new functionality for difficulty at various levels.
Earlier I had difficulty based on the level of recursion, but I
found that the change in difficulty wasn't noticeable.

So now what I do instead is that I will have numbers randomly
generated between 1 and 10 inclusive, and if those numbers fall
within a certain threshold, then I will generate a random move.
The threshold is 5 for the easy level, and 2 for the medium level.
So if you're playing on easy and the random number is <= 5, then
you generate a random move. Otherwise you run the minimax algorithm
to get the next move. The layers of recursion on the minimax
algorithm is 7 now.

To do:
1) Set the recursion levels in the GameVariables class in
   game_tracking_variables.py instead of hardcoding it

2) Consider refactoring the game functions in game_module.py