Refactored the code, fixed some bugs:
1) There were 2 separate functions to run the game when on single
player: one for when the computer went first, one for when the
user went first. This made for a lot of code duplication. I restructured
it such that it can just be taken care of in one function. This saved
a lot of code.

2) I found bugs in the code when the user signs in and out. There
was a situation where we would look in the list of people signed in
and just assume that they were string values. So now I check that
the contents are strings before calling casefold on it.