Fixed a bug in the new find_move function, updated .gitignore:
1) Bug fix:

Problem:

When playing on hard mode, there was an issue where you could find a
way to beat the game on hard mode by the following sequence (you're X):

	1) X| |   2) X|O|    3) X|O|   4) X|O|   5) X|O|
	   - - -     - - -      - - -     - - -     - - -
	    | |       | |       X| |      X| |      X|X|
	   - - -     - - -      - - -     - - -     - - -
	    | |       | |        | |      O| |      O| |

After 5), you would have a fork (2 unblocked rows of 2), so you have
an automatic win. This was the case if you had 3 layers of recursion on
the minimax algorithm or 5 layers. I ran a test on the minimax algorithm
(which I removed) when on 3 layers and when only the top left corner is
filled with an X by you, and the move the computer (O in the above
sequence) should have made at step 2) was:

	2) X| |
	   - - -
	    | |O
	   - - -
	    | |

After looking into the find_move function, I found that the issue was
with how the minimax algorithm was called. The person was given either
the player number 1 or 2, which would be used when filling
game_tracking_vars.instance_current_status_of_board_pieces. If the
person was given the number 1, the computer was given the number
2, and vice versa. These numbers are passed into the minimax function,
and the order in which they are passed into the the minimax function
determines whether the minimax function is making a move for you or
against you. So if you are the number 1 and you made the move at
step 1) (as X), if the minimax algorithm was called with player_number = 1
and other_player_number = 2, then the function would act as if it's next
move is for you and not against you. This is exactly what happened in
the above sequence. The only difference was that instead of putting down
and X, the algorithm put down an O. So the computer was accidentally
playing for you but putting down an O instead of an X in the above
example.

Solution:

I took into account whether the computer went first or second, and
set the player_num and other_player_num variables accordingly, to where
the computer is now working against you and not for you.

TL;DR - When the person went first on single-player mode, the find_move
function called minimax as if it was working for you instead of against
you, but just putting down the opposite symbol (depending on whether
you're X or O), leading you to believe the algorithm was making bad
moves.

2) Updated the .gitignore file so that changes in the virtual
environment "project_env" (installing of new dependencies) do not
show in the version history. Will add a requirement.txt file.