Added test for minimax, and added alpha beta pruning:
If you read to the commit message from i0a0589ad45790fffc641d52f7f28fb7255bd4141
you'll remember that there was an expected move for O to make
when you put down an X in the top left corner (below):

 X| |       X| |
 - - -      - - -
  | |   -->  | |O
 - - -      - - -
  | |        | |

I added in a test for this in the tests folder, and made sure that
alpha-beta pruning gave us the same move here. I also made sure
that alpha-beta pruning is used in the find_move function, used for
the computer to make a move when you play on single-player mode.