Initial design for the user interface:
This is the initial design for the user interface. I originally
tried doing this using the Tkinter library last year, however,
it proved to be very cumbersome, and I found the tkinter library
does not offer support for custom fonts. So I created the interface
using the pygame library. The text shown in the application uses
a custom font. The files for the custom font are represented by
the *.tff files.

The game_text.py module holds a class that represents text in the
game. The button.py odule holds a class that represents buttons
in the game. The application works by having a configuration file
representing each page, where settings are listed in the
configuration file. The application.py file is responsible for
loading the settings from one configuration file that will be
the main page, and then running the pygame loop that represents
the application. DO NOT alter the configuration files without
a thorough reading of the code. You will likely break the application
otherwise. A commit will follow with a "requirements.txt" file
listing the dependencies that the user must install for the
application to work.