Added code for the board and the game itself:
I added the code that makes up the Tic-Tac-Toe board,
and allows for either an X or an O to be drawn on
various pieces of the board. The code ensures that while
a figure is being drawn, no clicks by the user
on other pieces of the board result in any drawing. Once
a piece is finished being drawn on the board, then only
can another one be drawn. Things to implement:

1) The code must account for the usage of the minimax
algorithm. The program will use the minimax algorithm to
come up with a good move. With many layers of recursion,
this will be slow, and will interfere with the GUI we
create for the game. For this reason, the algorithm
will have to run asynchronously. Its returned result
will be used to draw the next piece. I am currently
learning about threads and processes to implement this.

2) Documentation must be added in tic_tac_toe_button.py